Tune location/orientation solver
This change should make LOSolver agnostic of the order of the points:
As no point should affect the calculation more than others, basis vector X (and indirectly Y) is calculated as follows (degrees used in this explanation):
"Imaginary" vectors from centroid, lying on the plane defined by points A, B and C and separated by 120 degrees (1/3 of full round) are rotated around the centroid so that the sum of the ("signed") angles between these and the vectors to real corners of the triangle are zero. "Imaginary" vector initially pointing to point A and rotated as explained above is then used as a basis vector X.

When generating videos (overlaid with the previous version) using LoSolverDemo_Godot, there's visible but insignificant difference. This is, however, more "correct" way to do this.