Fix writing&reading of lidar logs
Default size of float to log (when using <<-operator with QDataStream) was 64 bits instead of 32. This led to mismatch between written chunk size and the real number of bytes written.

Every day you learn something new. This one took several hours to find out (due to looking for the error from wrong places and (hopefully temporary) lack of ability to think clearly)...

Added also detection of duplicate time stamps as this was already done with other log types (which actually led to the discovery of the logging bug above).