Add TransformMatrixGenerator
- Added "scriptable" way to generate a transformation matrix. This allows usage of chained rotations using euler angles, translations and matrix multiplications to be used to generate a transformation matrix.

Added this mostly to be able to position a rotating lidar to a specified position in the antenna rig. This can, however also (and in fact after this commit only) be used to place anything in any position related to an arbitrary origin. This helps in situations like GNSS-antennas and camera attached to a vehicle. Using this you can place both in relation to common origin (previously if you wanted to solve the location/orientation of the camera you had to calculate the positions of the antennas in relation to the camera).

- Unified handling of transform matrices so that Eigen::Affine-type is used throughout. Hard to imagine situation when affine transforms won't suffice.

- Made some minor tunings to PostProcessingForm's UI and added descriptive labels.