Add unit tests (point filtering)
Finally took the time to figure out how to write unit tests in Qt (it was easy, as a real fan of unit testing should have done it far earlier). Difficulties testing the shallow slope filtering led to this.

And well, filtering seems to work, it just doesn't make point clouds much better.