Add RELPOSNED-"caching" to LOInterpolator
Thought that finding of RELPOSNEDs from QMap was somewhat costly operation and wrote a code to avoid doing it on every round. This caused speed-up of 1 (one!!1!) second out of 15 (with a sample size of 1, so even this is very questionable value). Likely wasn't worth it, but it's now done anyway...