Add option "Normal length as quality" to point cloud creation (lidar)
When activated, lengths of the normals written into xyz-file are multiplicative inverses of the measured lengths (by lidar). Some operations in MeshLab seemed to utilize this feature.

In reality the ambiguous xyz-file format should be replaced with ply, which allows f.ex quality to be written into a file as it's own field.