Fix some iterator operator "--"-issues
Defined behavior of "--"-operator on QMap's iterators is "Calling this function on QMap::begin() leads to undefined results" instead of going to .end() (which I expected). Not that I had ever stumbled upon these particular bugs (found this behavior when testing 80f60b2d042c8e6825e866c2d0add327a577a217)...

If this was "production quality"-code should write unit tests to find this kind of issues. But this is just tinkering, so who cares?

Not tested. It's possible that these (Stylus') functionalities will never be used again, but fixing anyway to prevent very hard to find problems in the future.