Changed interpolation of location/orientation to use quats (slerp)
While working on a video where the scanner was moving relatively fast (like 10 km/h), differences in interpolation of the LOSolver and lidarscript differed so that origins of the "laser beams" sometimes went outside the lidar unit. Using quat-slerp (or maybe even squads (but not now...)) is also the "right way" to do this, I just were too lazy to do it before stumbling upon problems.

This also affects generating of point clouds and raster cameras. Raster cameras still use the old method when using ITOW-timestamps. May not be a big issue since movement should be minimal when shooting stills.