Update Eigen to 3.4.0
Note: Vectorization still not working (issue #10 ) when compiling 32-bit (windows), shows "assertion failed" message box. This is actually why I tried to update to the newest stable version 3.4.0. This is hard to find problem since the assertion does not come up when compiling debug-version.

Maybe better to compile 32-bit versions without vectorization for now. You should really use 64-bit version for post-processing. But if using an ancient laptop with 32-bit OS for logging (like me) the 32-bit version is still needed.

BTW: Eigen from master-branch did not crash, but due to the nature of this problem this may be just luck so not relying on that.