Add preliminary lazy evaluator implementation and unit tests
Changed unit tests so that no separate executables are created (as by default if using the recommended "qt-way" (from some tutorials, that is)).

Wrote this "lazy evaluator" to be used for configurable point filters and to potentially speed those up. This is, however WIP for now.

This is more like a backup-commit. I didn't manage to get coverage analysis (using gcov, lcov) working in Windows, so switching to linux now to try there...