Add possibility to use margins with convex hulls
Margin is handled as shifts of faces of the convex hull in their normal's direction (positive = outward, negative = inward).

This style of scaling may cause hulls to "deform", especially when margin is negative. But this also should not cause much computational burden (almost free) and can be easily calculated real-time.

Use case for scaling could be for example picking points of a building in MeshLab and just feed those as convex hull points. Adding some positive margin to the hull then ensures that the whole building will be included when scanning.