UnitTests.pro: Force compiling c++17
This doesn't actually affect anything in Linux, where I'm currently
almost entirely working with this. Tested compiling in windows, however,
and it nagged that something needed c++17. So committing this from
Linux-side anyway. I was a bit worried about conflicts in pro-file if
committing/pushing from there (there was some uncommitted/pushed changes
here on the Linux-side when compiling in Windows).