Add partial/preliminary implementation of "PointFans"
Seems quite typical that when scanning buildings you may not get enough
points from the roof (because can't reach high enough with the scanner).
This causes Poisson surface reconstruction to generate very weird and
hard to fix shapes. Earlier (for the barn scan, for example) I just
wrote some code to generate points for roofs. But that's a quite
complicated so adding a way to generate points by defining "point fans"
instead.

This commit only adds partial implementation and the classes are only
used in unit tests for now. So this is more like a backup-commit.